Fuzz: Add immutable append_fuzzed_immut method to trait
NEXT: This allows to create a new type of FuzzChain[Immut] which
maintains immutable borrowed references to fuzzers. This can allow
the same fuzzer instance to be chained more than once in this new
type of FuzzChain[Immut]

NOTE: Most/All of the currently implemented fuzzers dont need to
change their contents, so adding a FuzzChainImmut should be fine.
Only LoopFixedStrings needs to know about which string in its list
to send next, but then it uses the step passed to append_fuzzed
to decide which should be next (NOTE: It also stores the same
in its internal curi field, but doesnt really use it in any
impactable way, so sum total everything should be fine).

All the currently implemented Fuzzers have been updated wrt this
new method in Fuzz trait.