Main: Add support for ioargs (HashMap, --ioarg k=v)
Allow user to pass arguments for the io bridge module.

It can be used for example by the TlsClient mode to set the domain
and as to whether server certificate should be verified or not.

Add a documentation for handle_cmdline, wrt the cmdline args
supported.

NOTE: the domain arg wrt tlsclient not captured in documentation,
bcas it will be changed to the --ioarg mechanism