ALERT: SingleSession multi request 10 and 1000 msec in btw
With 10msec each request seems to appear on client side

But 1000msec btw requests, only 1st request seems to get registered
on server side, but not subsequent requests within a single session.

TODO: Need to check with valid http requests (currently it is more
of invalid requests) if things are ok or not.