IOB:TlsClient:HttpSeperatePrg: Reading b4 shutdown still ...
Reading b4 shutdown to ensure there is no data, also doesnt help
with calling the shutdown 2nd time, now instead of the application
data on shutdown error, it is giving the syscall error.

So it would appear like the ssl wrappers in rust dont seem to handle
the possible use of 2 shutdown calls in the flow? Or is it that I
am glancing through the Ssl wrappers as well as openssl's documentation
wrongly? Need to read through the documentation properly later.

May be if a 2nd shutdown call after the 1st returns ApplicationData
OnShutdown error, then read bufs and discard. And either way try
call read_till_end once again to check if we get a 0 value, to
indicate EOF, and be done with it.