VM:DataM:Compile: Simplify and cleanup the logic and end use
Rather now anything

that starts with a numeric char or + or - will be treated has a
numeric literal.

If it starts or ends with double quotes, it will be treated as
a string literal.

If it starts with $0x then it will be treated has a binary buffer
specified has a hex string.

If it starts with __ then it will be treated has a special value.

If none of above, then it will be treated as a var name. Additionally
it should start with a alphabhet.