VM:DataM:GetHelpers: BruteForce varName through possible argMapping
Currently, all DataM Get helper calls involving var names, will be
cross checked to see, if the var name is a function arg by running
it through arg mapping.

TODO: Later optimise this path, by having the compile time set a
flag wrt DataM entity created to indicate whether it is a global
var or a func arg. So that only if it is a func arg, this run
through arg map is done.