Tests:IntroPRG: Reuse local var names in a destructive but ok way
After the local variables have been set in 1st function, it calls
the 2nd function, which assigns new set of values to its local
variables, which have same name as the local vars in 1st func.

Inturn on returns back to 1st function, print the local vars in
1st function.

If local vars stacking is not handled properly, either the 1st
or 2nd func will print wrong values.