Tests:LetLocal:Src autotype infering known corner case triggering.
Bcas of how it was implemented previously, and its assumptions,
checking for type of a func arg, appearing as the src entity
in a letlocal, will trigger a panic.

Previously farg2real was expected to be called only at runtime,
as it will have its reqd info only then. So also local and global
var lists (rather hm).