Tests:IntroPrg: Update wrt new support to pass local var to func
And inturn it being passed into additional child functions.

FUTURE: The mechanism implemented is flexible and can allow such
passed func args to be modified, if required in future, which inturn
will allow the modified value to be used by any of the parent funcs
also, ie till the parent function, which inturn declares the local
variable, that was passed around.