VM:Op:CallX: Allow vars and literals to be passed as func args
Add a PreOps vector to Context. This allows any Opcode/Instruction
to generate a set of additional instructions to be inserted before
the current instruction, in the opcodes vector.

This inturn is used by a new CallX pseudo instruction to allow
even literal values to be passed as function arguments.

Inturn to help with this, All Op and related data types have been
extended to support Clone operation using auto derive logic.