Readme: Cleanup including renaming prg file asm script file
Cleanup things in general

Rename the fuzzerk's script file as asm script file rather than
as a program file, to avoid confusion with 3rd party programs that
can be tested using fuzzerk.

Update to use the new syntax of $0x rather than 0x wrt hex string
in cfg files (this was done to make it similar to that supported
in asm script files).