FuzzChains: Remove FuzzChainImmut and switch to FuzzChain
As of now, the fuzzer can decide, whether it should be called in
a mutable state or immutable state, when fuzzed data is being
generated from a fuzzchain.

FuzzChainImmute could be useful, in that, if trying to debug any
issue with custom fuzzers and its users, then by making it immutable
throught a fcimmut one could potentially reduce the search space for
debugging in some scenarios. However the same can be obtained
indirectly by not implementing/overriding need_mutable method of
fuzzer trait.