VM:Op:Jump+:P1: Rename existing Jump to JumpRaw, Add new Jump
This is part of adding a 2 step compile, to avoid trying to find
the label to jump to at runtime, instead of it fix up the addr
to jump as part of a 2nd pass.

However the old Jump ie JumpRaw now, still remains, for now, incase
I want to use it somewhere, without trying to run 2nd compile phase.